% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdc_infant_death}
\alias{cdc_infant_death}
\title{Infant births and deaths in the US}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12120 rows and 6 columns.}
\source{
CDC
}
\usage{
cdc_infant_death
}
\description{
A data set
}
\details{
These data are based on
}
\examples{
data(cdc_infant_death, package="graphclassmate")
cdc_infant_death 
}
\keyword{datasets}
