% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_income}
\alias{county_income}
\title{Median income by US county in 2016}
\format{A tidy data frame (tibble) with 816 observations and 4 variables.
    
\describe{
  \item{county_id}{US counties encoded by a 5-digit number (FIPS code) which 
  uniquely identifies 816 counties and county equivalents}
  \item{county}{Name of the county or county equivalent}
  \item{state}{US states and the District of Columbia} 
  \item{income}{Median income over the past 12 months}
  }}
\source{
ACS Table B07011, ACS_05_EST_B07011
}
\usage{
county_income
}
\description{
American Community Survey (ACS) median income in the past 12 months 
in constant 2016 dollars.
}
\details{
Source: ACS Table B07011 "Median income in the past 12 months by 
geographic mobility in the past year for current residence in the US." 
Population 15 years and older in the US with income.
}
\examples{
data(county_income, package = "graphclassmate")
county_income 
}
\keyword{datasets}
