% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linked_birth_infant_death}
\alias{linked_birth_infant_death}
\title{Infant births and deaths in the US, 2007-2016}
\format{A tidy data frame (tibble) with 12,120 observations and 6 variables.
    
\describe{
  \item{region}{Geographical regions of the US as denoted by the federal 
  census: CENS-R1 (Northeast), CENS-R2 (Midwest), CENS-R3 (South), 
  and CENS-R4 (West).}
  \item{county}{US counties encoded by a 5-digit number (FIPS code) which 
  uniquely identified 303 counties and county equivalents.}
  \item{race}{Race or ethnicity: Amerind, Asian/PI, Black, Hispanic, White.}
  \item{age}{Age group of the mother: 15, 15-19, 20-24, 25-29, 30-34, 35-39, 
  40-44, 45-49. These bins were selected by the CDC.}
  \item{deaths}{The number of deaths of infants less than a year old, 
  2007--2016.}
  \item{births}{The number of births, 2007--2016.}
  }}
\source{
US Centers for Disease Control and Prevention (CDC) WONDER online 
    database, https://wonder.cdc.gov/controller/datarequest/D69
}
\usage{
linked_birth_infant_death
}
\description{
CDC linked birth and infant death records in the US from 2007 to 2016.
}
\details{
The CDC data set includes US counties encoded by the 5-digit Federal 
Information Processing Specification (FIPS) county codes--codes that were 
superseded in 2009 by the INCITS 31 codes. County-level data are shown only 
for counties with populations of 250,000 or more. Because of population 
changes, data is not available for all counties in all years.

Race/ethnicity data are grouped and summarized into 5 levels: Amerind for 
indigenous peoples excluding Native Hawaiians, Asian/PI for people of Asian 
and Pacific Islander descent, Black for African Americans, Hispanic for 
people of Latin American descent, and White.
}
\examples{
data(linked_birth_infant_death, package = "graphclassmate")
linked_birth_infant_death 
}
\keyword{datasets}
