% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metro_pop}
\alias{metro_pop}
\title{Population in the NY metro area}
\format{A tidy data frame (tibble) with 60 observations and 3 variables.
    An observation is the population in a county by race/ethnicity.  
    
\describe{
  \item{race}{Race or ethnicity}
  \item{county}{Name of county}
  \item{population}{Number of residents from the 2000 US census}
}}
\source{
\code{r-graph-catalog} by Joanna Zhao and Jenny Bryan, https://github.com/jennybc/r-graph-catalog
}
\usage{
metro_pop
}
\description{
A data set of population in the New York metropolitan area by county and 
race/ethnicity from the 2000 census.
}
\details{
These data are taken from section 8.5 in Robbins (2013).
}
\examples{
data(metro_pop, package = "graphclassmate")
metro_pop
}
\keyword{datasets}
