% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{museum_exhibits}
\alias{museum_exhibits}
\title{Visit duration at museum exhibitions}
\format{A tidy data frame (tibble) with 588 observations and 2 variables.
    An observation is a unique person at an exhibit. 
    
\describe{
  \item{exhibit}{Name of the exhibition}
  \item{minutes}{Number of minutes a unique person spent viewing the exhibit}
}}
\source{
\code{r-graph-catalog} by Joanna Zhao and Jenny Bryan, https://github.com/jennybc/r-graph-catalog
}
\usage{
museum_exhibits
}
\description{
A data set of how much time visitors spend viewing museum exhibits.
}
\details{
These data are simulated to approximate the data shown in Figure 4.10 in 
Robbins (2013). The original data were not readily available.
}
\examples{
data(museum_exhibits, package="graphclassmate")
museum_exhibits
}
\keyword{datasets}
