% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nontraditional}
\alias{nontraditional}
\title{Nontraditional and traditional undergraduates}
\format{A tidy data frame (tibble) with 269,057 observations and 5 variables.
    An observation is a unique student. 
    
\describe{
  \item{sex}{Student sex (Female or Male)}
  \item{race}{Student race or ethnicity (Asian, Black, Hispanic, or White)}
  \item{path}{Student path through curriculum (Nontraditional or Traditional)}
  \item{SAT}{Student SAT score at matriculation}
  \item{enrolled}{Number of years student enrolled at the institution from which they graduate}
}}
\source{
MIDFIELD https://engineering.purdue.edu/MIDFIELD
}
\usage{
nontraditional
}
\description{
A data set of the number of years that traditional and nontraditional 
students were enrolled at the US institutions from which they graduated.
}
\details{
These data are subset of MIDFIELD data. All students were enrolled at least 
one term in an engineering program. All students graduated though not 
necessarily in engineering.

The data subset excludes the 10th and 90th deciles of years enrolled.
}
\examples{
data(nontraditional, package = "graphclassmate")
nontraditional
}
\keyword{datasets}
