% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucb_admit}
\alias{ucb_admit}
\title{Student admissions at UC Berkeley}
\format{A tidy data frame (tibble) with 12 observations and 4 variables.
    An observation, by sex and department, records the number of people  
    who applied and the number admitted. 
    
\describe{
  \item{sex}{Sex of applicant, Female or Male}
  \item{dept}{Department A through F}
  \item{admitted}{Number admitted to the department}
  \item{applied}{Number of applicants to the department}
}}
\source{
Base R.
}
\usage{
ucb_admit
}
\description{
A data set on applicants to graduate school at University of California, 
Berkeley, for the six largest departments in 1973.
}
\details{
These data are based on the \code{UCBAdmissions} data in base R. The new 
variable \code{applied} is the sum of the number of students admitted and 
the number of students rejected by each department.
}
\examples{
data(ucb_admit, package = "graphclassmate")
ucb_admit 
}
\keyword{datasets}
