% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphclassmate.R
\docType{package}
\name{graphclassmate}
\alias{graphclassmate}
\alias{graphclassmate-package}
\title{Companion materials for a course in data visualization}
\description{
A collection of data sets to illustrate different types of graphs plus an 
opinionated ggplot2 theme.
}
\details{
\itemize{
  \item \code{theme_graphclass()} is a slightly modified version of 
    \code{theme_minimal()} from \pkg{ggplot2} 
  \item \code{rcb()} provides selected hex color codes from \pkg{RColorBrewer} 
}
}
\section{Data}{


The data samples from various sources provide a set of test cases for 
tutorials and exercises for commonly encountered graph types. 

\itemize{
  \item \code{rcb_colors} Selected named colors from the ColorBrewer palettes 
  \item \code{speed_ski} A subset of skiing data from \pkg{GDAdata} 
}
}

