% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speed_ski}
\alias{speed_ski}
\title{Speed skiing data}
\format{A tidy data frame (tibble) with 91 observations and 3 variables.
    An observation is a unique person in an event.
    
\describe{
  \item{event}{One of three downhill events: Speed Downhill, Speed Downhill 
  Junior, and Speed One.}
  \item{sex}{Female or male.}
  \item{speed}{Speed in km/hr.}
}}
\source{
GDAdata package https://CRAN.R-project.org/package=GDAdata
}
\usage{
speed_ski
}
\description{
A dataset of the speed of skiers from the downhill events in the 2011 world
championships.
}
\details{
These data are a subset of the \code{SpeedSki} dataset in the GDAdata 
package, arranged for ease of use in demonstrating graphs for 
univariate data.
}
\examples{
data(speed_ski, package="graphclassmate")
speed_ski
}
\keyword{datasets}
